Move check_table stuff to CheckPlugin
CheckPlugin uses the few remaining types defined there.  This also
fixes circ imports.

CMK-12298
