+++ cmk/base/api/agent_based/cluster_mode.py
@@ -11,7 +11,7 @@ from typing import Any, Final, Literal, NamedTuple, Protocol
 
 from cmk.utils.hostaddress import HostName
 
-from cmk.checkengine.check_table import ServiceID
+from cmk.checkengine.checking import ServiceID
 from cmk.checkengine.checkresults import state_markers
 
 from cmk.base.api.agent_based.checking_classes import (
