+++ cmk/base/core_config.py
@@ -28,8 +28,7 @@ from cmk.utils.paths import core_helper_config_dir
 from cmk.utils.servicename import ServiceName
 from cmk.utils.store import load_object_from_file, lock_checkmk_configuration, save_object_to_file
 
-from cmk.checkengine.check_table import ConfiguredService, ServiceID
-from cmk.checkengine.checking import CheckPluginName, Item
+from cmk.checkengine.checking import CheckPluginName, ConfiguredService, Item, ServiceID
 from cmk.checkengine.parameters import TimespecificParameters
 
 import cmk.base.api.agent_based.register as agent_based_register
