+++ cmk/checkengine/checking/__init__.py
@@ -4,7 +4,7 @@
 # conditions defined in the file COPYING, which is part of this source code package.
 
 from ._checking import CheckPluginName, CheckPluginNameStr, Item
-from ._plugin import AggregatedResult, CheckPlugin
+from ._plugin import AggregatedResult, CheckPlugin, ConfiguredService, ServiceID
 from ._timing import make_timing_results
 
 __all__ = [
@@ -12,6 +12,8 @@ __all__ = [
     "CheckPlugin",
     "CheckPluginName",
     "CheckPluginNameStr",
+    "ConfiguredService",
     "Item",
     "make_timing_results",
+    "ServiceID",
 ]
