+++ tests/unit/cmk/base/agent_based/checking/test_cluster_modes.py
@@ -11,8 +11,7 @@ import pytest
 
 from cmk.utils.hostaddress import HostName
 
-from cmk.checkengine.check_table import ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, ServiceID
 
 from cmk.base.api.agent_based import cluster_mode
 from cmk.base.api.agent_based.checking_classes import (
