+++ tests/unit/cmk/base/api/agent_based/value_store/test_vs_global_state.py
@@ -8,8 +8,7 @@ from pytest import MonkeyPatch
 import cmk.utils.store as store
 from cmk.utils.hostaddress import HostName
 
-from cmk.checkengine.check_table import ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, ServiceID
 
 from cmk.base.api.agent_based.value_store._global_state import (
     get_value_store,
