+++ tests/unit/cmk/base/api/agent_based/value_store/test_vs_utils.py
@@ -12,8 +12,7 @@ import pytest
 from cmk.utils import store
 from cmk.utils.hostaddress import HostName
 
-from cmk.checkengine.check_table import ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, ServiceID
 
 from cmk.base.api.agent_based.value_store._utils import (
     _DiskSyncedMapping,
