+++ tests/unit/cmk/base/test_check_table.py
@@ -14,8 +14,7 @@ from cmk.utils.hostaddress import HostName
 from cmk.utils.rulesets import RuleSetName
 from cmk.utils.tags import TagGroupID, TagID
 
-from cmk.checkengine.check_table import ConfiguredService, ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, ConfiguredService, ServiceID
 from cmk.checkengine.discovery import AutocheckEntry
 from cmk.checkengine.legacy import LegacyCheckParameters
 from cmk.checkengine.parameters import TimespecificParameters, TimespecificParameterSet
